renamed the folder for the camera properties created node editor for the camera component added to/from json code in the camera component added camera component to the ui shelf added camera prefab
