+++ packages/ui/src/components/editor/properties/cameraProperties/index.stories.tsx
